requirements.txt
cohere>=4.0.0
llama-index>=0.9.0
llama-index-embeddings-cohere>=0.1.0
llama-index-llms-cohere>=0.1.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
streamlit>=1.28.0